+++ src/love/mouse.h
@@ -1,9 +1,9 @@
 #ifndef SRC_LOVE_MOUSE_H_
 #define SRC_LOVE_MOUSE_H_
 
-#include <SDL.h>
 #include <string>
 #include "Types/Graphics/Point.h"
+#include "libretro.h"
 
 namespace love {
 
@@ -13,43 +13,6 @@ namespace love {
 class mouse {
 	public:
 	bool load();
-	/**
-	 * @brief Sets the current visibility of the cursor.
-	 *
-	 * @param visible True to set the cursor to visible, false to hide the cursor.
-	 *
-	 * @see love.mouse.getVisible
-	 */
-	mouse& setVisible(bool visible);
-
-	/**
-	 * @brief Checks if the cursor is visible.
-	 *
-	 * @return True if the cursor to visible, false if the cursor is hidden.
-	 *
-	 * @see love.mouse.setVisible
-	 */
-	bool isVisible();
-
-	/**
-	 * @brief Sets the current X position of the mouse.
-	 *
-	 * @param x The new position of the mouse along the x-axis.
-	 *
-	 * @return The mouse module.
-	 */
-	mouse& setX(int x);
-
-	/**
-	 * @brief Sets the current Y position of the mouse.
-	 *
-	 * @param y The new position of the mouse along the y-axis.
-	 *
-	 * @return The mouse module.
-	 *
-	 * @see love.mouse.getY
-	 */
-	mouse& setY(int y);
 
 	/**
 	 * @brief Returns the current x-position of the mouse.
@@ -92,14 +55,17 @@ class mouse {
 	 */
 	bool isDown(const std::string& button);
 
+	void update();
+
 	int getButtonKey(const std::string& button);
 	std::string getButtonName(int button);
 
-	int m_x, m_y;
-	int buttonState[20];
+	int16_t m_x, m_y;
+	int16_t buttonState[RETRO_DEVICE_ID_MOUSE_BUTTON_5];
 
-	void moveEvent(SDL_MouseMotionEvent event);
-	void buttonEvent(SDL_MouseButtonEvent event);
+	void mousemoved(int x, int y, int dx, int dy);
+	void mousepressed(int x, int y, const std::string& button);
+	void mousereleased(int x, int y, const std::string& button);
 };
 
 }  // namespace love