+++ src/love/script.cpp
@@ -242,10 +242,6 @@ script::script(const std::string& file) {
 	chai.add(fun(&system::getVersionString), "getVersionString");
 
 	// Mouse
-	chai.add(fun(&mouse::setVisible), "setVisible");
-	chai.add(fun(&mouse::isVisible), "isVisible");
-	chai.add(fun(&mouse::setX), "setX");
-	chai.add(fun(&mouse::setY), "setY");
 	chai.add(fun(&mouse::getX), "getX");
 	chai.add(fun(&mouse::getY), "getY");
 	chai.add(fun(&mouse::getPosition), "getPosition");
@@ -363,21 +359,21 @@ script::script(const std::string& file) {
 		hasjoystickreleased = false;
 	}
 	try {
-		chaimousepressed = chai.eval<std::function<void(int, int, int)> >("mousepressed");
+		chaimousepressed = chai.eval<std::function<void(int, int, const std::string&)> >("mousepressed");
 	}
 	catch (const std::exception& e) {
 		std::cout << "[ChaiLove] [script] mousepressed() " << e.what() << std::endl;
 		hasmousepressed = false;
 	}
 	try {
-		chaimousereleased = chai.eval<std::function<void(int, int, int)> >("mousereleased");
+		chaimousereleased = chai.eval<std::function<void(int, int, const std::string&)> >("mousereleased");
 	}
 	catch (const std::exception& e) {
 		std::cout << "[ChaiLove] [script] mousereleased() " << e.what() << std::endl;
 		hasmousereleased = false;
 	}
 	try {
-		chaimousemoved = chai.eval<std::function<void(int, int)> >("mousemoved");
+		chaimousemoved = chai.eval<std::function<void(int, int, int, int)> >("mousemoved");
 	}
 	catch (const std::exception& e) {
 		std::cout << "[ChaiLove] [script] mousemoved() " << e.what() << std::endl;
@@ -514,7 +510,7 @@ void script::joystickreleased(int joystick, const std::string& button) {
 	#endif
 }
 
-void script::mousepressed(int x, int y, int button) {
+void script::mousepressed(int x, int y, const std::string& button) {
 	#ifdef __HAVE_CHAISCRIPT__
 	if (hasmousepressed) {
 		try {
@@ -528,7 +524,7 @@ void script::mousepressed(int x, int y, int button) {
 	#endif
 }
 
-void script::mousereleased(int x, int y, int button) {
+void script::mousereleased(int x, int y, const std::string& button) {
 	#ifdef __HAVE_CHAISCRIPT__
 	if (hasmousereleased) {
 		try {
@@ -542,11 +538,11 @@ void script::mousereleased(int x, int y, int button) {
 	#endif
 }
 
-void script::mousemoved(int x, int y) {
+void script::mousemoved(int x, int y, int dx, int dy) {
 	#ifdef __HAVE_CHAISCRIPT__
 	if (hasmousemoved) {
 		try {
-			chaimousemoved(x, y);
+			chaimousemoved(x, y, dx, dy);
 		}
 		catch (const std::exception& e) {
 			std::cout << "[ChaiLove] [script] Failed to call mousemoved(): " << e.what() << std::endl;